feat: #id 12001 Allow preload or inject of windowTitleBar
Modified windowTitleBar component so that it can be preloaded/injection
Added WindowManager.deliveryMechanism under application config.
